REFACTOR #139 side-effect free level parsing after upload
